requirements.txt
Flask
Flask-SQLAlchemy
pytesseract
datetime
pillow
PyMySQL
